requirements.txt
click==6.7
Flask==0.12.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
PyYAML==3.12
Werkzeug==0.12.1
